Bug 2006125 - Clean up helper functions for custom stylelint rules r=desktop-theme-reviewers,mstriemer
Differential Revision: https://phabricator.services.mozilla.com/D284800
